requirements.txt
aiosmtpd==1.4.4.post2
asgiref==3.7.2
atpublic==4.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
PyYAML==6.0
pyyml==0.0.2
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3